requirements.txt
django==4.1.11
django-htmx==1.16.0
djongo==1.3.6
fastapi==0.103.1
pydantic==1.10.12
pyppeteer==1.0.2
uvicorn==0.23.2
whitenoise==6.5.0
python-decouple==3.8
pymongo==3.12.3
fontawesomefree==6.4.2
requests==2.29.0
httpx==0.25.0
openai==1.1.1
pinecone-client==2.2.4
htmlmin==0.1.12
playwright
pandas
